NULL check for HelloWorldClient proxy
I ran into this a few times with older versions. It can happen if
there's a version mismatch but I seemed to get it also after rebuilding
all with one version.  In any case, it went away with 3.1.10. This is
just some code to write a message instead of segfaulting on the pointer
dereference.

Other than that, I've tried to do minimal changes (so far) to the
example as it is written in Wiki, only encoding the steps in the script.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>