Update all runtime & tools to latest 3.1.12.x
It's all done in one go here because it involves a change to use
different versions per package and I don't have time to split the
commits now...

Versions:

CORE_TOOLS_VERSION=3.1.12.3
DBUS_TOOLS_VERSION=3.1.12.1
SOMEIP_TOOLS_VERSION=3.1.12.1
CORE_RUNTIME_VERSION=3.1.12.4
DBUS_RUNTIME_VERSION=3.1.12.7
SOMEIP_RUNTIME_VERSION=3.1.12.12

When the script was created, the versions were mostly in sync and since
it is just an example it was not that important.  Nowadays, there are
different patchreleases for each variant, so the versions should be
treated separately.  Of course you need to use mutually-compatible
versions of both tools and runtime.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>