Don't fail if patch is applied already
It's common to re-run the script in an already patched source tree.
patch will fail (and script will pause) if it's already applied, unless
we take care of it.

Another option would be to reset the appropriate source trees, but
perhaps the user wants to make and keep local changes, so this seems
better.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>