Support running on systems without sudo
The program can be run without sudo if it is started with the root user.
Previously it would just fail, even if run with the root user.

sudo might not be installed (see below).

Some additional package installations slipped through here also:
(unzip, pkg-config, etc.)  These are things that may not be installed on
minimal distros such as docker base images.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>