Add local install dir to find boost libraries
Cmake should be instructed to compile the generated code with a library
search path that includes the place where boost was installed.

This may have worked in the past because of a system-installed boost, or
vsomeip libraries may have included path references that made the
library be found anyway, or it may be compiler version dependent
somehow.  Anyhow, this seems to solve the problem.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>